feat(unicorn): use warning for esm rules
As much as I love Sindre's push for ESM in the community, there are still instances where ESM is not working at this moment. This module is one of those instances, due to ESLint not supporting ESM sharable configs. Using warnings for ESM rules will ensure that we can continue pushing the ecosystem forward, while not outright penalizing modules for their environments.